SPARKC-96: Provide more fault-tolerant batch mutations
saveToCouchbase currently tries to write documents in batches of 128
(the default for Observable parallelism).

Results from the field, and my own testing, indicate that this can be too
much and lead to timeouts.

This change adds an optional maxConcurrent parameter to limit the batch
size.  I was tempted to reduce the default batch size too, but stayed
clear of that for fear of slowing existing applications that can handle
128 fine.

Change-Id: I394e579b3a4d2aaeebb955b0199390e8626e8e7a
Reviewed-on: http://review.couchbase.org/109572
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Graham Pople <grahampople@gmail.com>